Tubes_UTS/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
mysqlclient==2.2.4
python-dotenv==1.0.1
requests==2.31.0
